Add sys.global slmctl control for global mount
git-svn-id: svn+ssh://frodo/cluster/svn/projects/slash_nara@25258 3eda493b-6a19-0410-b2e0-ec8ea4dd8fda

+++ slashd/rmc.c
@@ -68,7 +68,7 @@
 #include "lib/libsolkerncompat/include/errno_compat.h"
 #include "zfs-fuse/zfs_slashlib.h"
 
-int			use_global_mount;
+long			use_global_mount;
 
 uint64_t		slm_next_fid = UINT64_MAX;
 psc_spinlock_t		slm_fid_lock = SPINLOCK_INIT;
